git-svn-id: http://svn.seqan.de/seqan/trunk@4452 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/seqcons/seqcons.cpp
@@ -438,7 +438,9 @@ int main(int argc, const char *argv[]) {
 	for(TSize currentContig = 0; currentContig < numberOfContigs; ++currentContig) {
 
 		if (consOpt.realign) {
-			Score<double, WeightedConsensusScore<Score<double, FractionalScore>, Score<int, ConsensusScore> > > combinedScore;
+			Score<int, WeightedConsensusScore<Score<int, FractionalScore>, Score<int, ConsensusScore> > > combinedScore;
+			//Score<int, ConsensusScore> combinedScore;
+			//Score<int, FractionalScore> combinedScore;
 			reAlign(fragStore, combinedScore, currentContig, consOpt.bandwidth);
 		} else {
